+++ front/pages/_app.tsx
@@ -6,6 +6,8 @@ import Link from "next/link";
 import { SessionProvider } from "next-auth/react";
 import { MouseEvent, ReactNode } from "react";
 
+import { NotificationArea } from "@app/components/sparkle/Notification";
+
 function NextLinkWrapper({
   href,
   className,
@@ -49,7 +51,9 @@ export default function App({
   return (
     <SparkleContext.Provider value={{ components: { link: NextLinkWrapper } }}>
       <SessionProvider session={session}>
-        <Component {...pageProps} />
+        <NotificationArea>
+          <Component {...pageProps} />
+        </NotificationArea>
       </SessionProvider>
     </SparkleContext.Provider>
   );
